cmdb/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-modeladmin-reorder==0.3.1
django-otp==0.9.1
django-reversion==3.0.7
django-split-settings==1.0.1
djangorestframework==3.11.0
drf-yasg==1.17.1
elfinder==0.0.3
elfinder-client==2.1.55a6
gmssl==3.2.1
hupper==1.10.2
idna==2.9
imjoy-elfinder==0.1.55
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
PasteDeploy==2.1.0
pathvalidate==2.3.0
Pillow==7.1.2
plaster==1.0
plaster-pastedeploy==0.7
pyparsing==2.4.7
pyramid==1.10.4
pyramid-jinja2==2.8
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
qrcode==6.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
translationstring==1.3
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
venusian==3.0.0
waitress==1.4.4
WebOb==1.8.6
Werkzeug==1.0.1
zope.deprecation==4.4.0
zope.interface==5.1.0
